% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexplore.R
\name{Chisq}
\alias{Chisq}
\title{Chisq - chi-square and c-stats on non-binned basis}
\usage{
Chisq(df, Y, Vars)
}
\arguments{
\item{df}{Dataframe}

\item{Y}{Binary event variable (0, 1) name. Integer (\code{Int}) is required.}

\item{Vars}{Dataframe of the variable summary table from function Vars().}
}
\value{
Dataframe of variable chi-square summary.
}
\description{
Chisq - chi-square and c-stats on non-binned basis
}

