% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexplore.R
\name{Freq}
\alias{Freq}
\title{Categorical Variable Frequency Table}
\usage{
Freq(df, Vars)
}
\arguments{
\item{df}{Dataframe}

\item{Vars}{Dataframe of Variable Summary Table from function Vars()}
}
\value{
Dataframe of frequency table
}
\description{
Categorical Variable Frequency Table
}

