% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexplore.R
\name{Stats}
\alias{Stats}
\title{Statistical Analysis Summary}
\usage{
Stats(df, Y, Vars)
}
\arguments{
\item{df}{Dataframe}

\item{Y}{Binary event variable (0, 1) name. Integer (\code{Int}) is required.}

\item{Vars}{Dataframe of Variable Summary Table from function Vars()}
}
\value{
Dataframe of statistical summary
}
\description{
Statistical Analysis Summary
}

