% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexplore.R
\name{Univars}
\alias{Univars}
\title{Univariate Analysis}
\usage{
Univars(df, Vars)
}
\arguments{
\item{df}{Dataframe}

\item{Vars}{Dataframe of variable summary table from function Vars()}
}
\value{
Dataframe of univariate summary
}
\description{
Univariate Analysis
}

