% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexplore.R
\name{Varclus}
\alias{Varclus}
\title{Variable Clustering Analysis}
\usage{
Varclus(df, Y, Vars, maxEigenValue = 0.7)
}
\arguments{
\item{df}{Dataframe}

\item{Y}{Binary event variable (0, 1) name. Integer (\code{Int}) is required.}

\item{maxEigenValue}{Eigen value cutoff, defualt = 0.7.}
}
\value{
Dataframe of Variable Cluster Analysis results.
}
\description{
Variable Clustering Analysis
}

