% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexplore.R
\name{varexplore}
\alias{varexplore}
\title{Building a set of variable explore tools mainly for binary prediction.}
\usage{
varexplore(df, Y, groups = 20)
}
\arguments{
\item{df}{Dataframe}

\item{Y}{Binary event variable (0, 1) name. Integer (\code{Int}) is required.}

\item{groups}{Integer variable for default grouping in numeric variable.
Default is 20 groups / bins}
}
\value{
Generate a Shiny Dashboard object containing 5 tabs: Variable
Summary, Variable Clustering, Log-Odds Charts for numeric variables
interactive binning, Log-Odds Charts for categorical variables, and export
results to CSV file format
}
\description{
Output into Shiny Dashboard for user interaction, especially numeric
variable binning and WOE summary with SQL codes available for transformation.
}

