% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexplore.R
\name{Vars}
\alias{Vars}
\title{Create Variable Summary Table}
\usage{
Vars(df, sv = c(9999, 99999), maxCatLevel = 50, maxNumPct = 0.99)
}
\arguments{
\item{df}{Dataframe}

\item{sv}{Vector of values user defined and can be detected if unusually
high proportion of the variable has the values; default = c(9999, 99999).}

\item{maxCatLevel}{Maximum categorical levels to be excluded in Shiny Dashboard
output; default = 50.}

\item{maxNumPct}{Maximum percent of one single value of the variable to be
exlcuded in Shiny Dashboard.}
}
\value{
Dataframe of variable summary table.
Variable Reduction Rules:
  1. Categorical and more than maxCatLevel (default) 50
  2. Only one value
  3. Numeric variable & one value more than (default) 99%
  4. Missing value more than (default) 99%
  5. Zero value more than 99.5%
  6. Outlier (over 5 standard deviation) more than 99.5%
  7. Special value more than 99.5%
}
\description{
Create Variable Summary Table with general rules to drop
  certain variables with low value to model / explore. Act as a primary
  variable selection for the rest of the package functions
}

